Clean up StrongNameFileSystem API
This API is meant as a file system abstraction that can be mocked /
replaced during testing. Reduced the API to the necessary member and
moved strong name logic to the appropriate type.
