+++ src/Compilers/Core/Portable/StrongName/StrongNameFileSystem.cs
@@ -7,6 +7,9 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis
 {
+    /// <summary>
+    /// This is an abstraction over the file system which allows for us to do more thorough unit testing.
+    /// </summary>
     internal class StrongNameFileSystem
     {
         internal readonly static StrongNameFileSystem Instance = new StrongNameFileSystem();
@@ -18,40 +21,6 @@ namespace Microsoft.CodeAnalysis
         {
             Debug.Assert(PathUtilities.IsAbsolute(fullPath));
             return File.ReadAllBytes(fullPath);
-        }
-
-        /// <summary>
-        /// Resolves assembly strong name key file path.
-        /// </summary>
-        /// <returns>Normalized key file path or null if not found.</returns>
-        internal string ResolveStrongNameKeyFile(string path, ImmutableArray<string> keyFileSearchPaths)
-        {
-            // Dev11: key path is simply appended to the search paths, even if it starts with the current (parent) directory ("." or "..").
-            // This is different from PathUtilities.ResolveRelativePath.
-
-            if (PathUtilities.IsAbsolute(path))
-            {
-                if (FileExists(path))
-                {
-                    return FileUtilities.TryNormalizeAbsolutePath(path);
-                }
-
-                return path;
-            }
-
-            foreach (var searchPath in keyFileSearchPaths)
-            {
-                string combinedPath = PathUtilities.CombineAbsoluteAndRelativePaths(searchPath, path);
-
-                Debug.Assert(combinedPath == null || PathUtilities.IsAbsolute(combinedPath));
-
-                if (FileExists(combinedPath))
-                {
-                    return FileUtilities.TryNormalizeAbsolutePath(combinedPath);
-                }
-            }
-
-            return null;
         }
 
         internal virtual bool FileExists(string fullPath)
